+++ app/Http/Controllers/ResetPasswordController.php
@@ -2,12 +2,10 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Password;
-use Illuminate\Auth\Events\PasswordReset;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Str;
 use App\Http\Requests\User\ResetRequest;
+use App\Services\ResetPasswordService;
+
+use Illuminate\Http\Request;
 
 class ResetPasswordController extends Controller
 {
@@ -28,26 +26,11 @@ class ResetPasswordController extends Controller
    * パスワード再設定更新
    *
    * @param  \App\Http\Requests\User\ResetRequest  $request
+   * @param  \App\Services\ResetPasswordService $resetPasswordService
    * @return \Illuminate\Http\Response
    */
-  public function update(ResetRequest $request)
+  public function update(ResetRequest $request, ResetPasswordService $resetPasswordService)
   {
-
-    $status = Password::reset(
-      $request->only('email', 'password', 'password_confirmation', 'token'),
-      function ($user, $password) {
-        $user->forceFill([
-          'password' => Hash::make($password)
-        ])->setRememberToken(Str::random(60));
-
-        $user->save();
-
-        event(new PasswordReset($user));
-      }
-    );
-
-    return $status === Password::PASSWORD_RESET
-      ? redirect()->route('mypage.login')->with('status', __($status))
-      : back()->withErrors(['email' => [__($status)]]);
+    return $resetPasswordService->updatePassword($request);
   }
 }
